Fix error handling in transaction page
In case of a not found it was fine to keep retrying for 10 times to get
the transaction as it could just be not yet synced by the
explorer-backend or it's in between the mempool and the confirmed state.

If we receive any other error it's not worth retrying as it will always
we wrong.